+++ services/__init__.py
@@ -2,5 +2,11 @@
 from .vault_stream_crawler import VaultStreamCrawler
 from .graph_query_engine import GraphQueryEngine
 from .incremental_tracker import CrawlStateTracker
+from .content_index import ContentIndex
 
-__all__ = ["VaultStreamCrawler", "GraphQueryEngine", "CrawlStateTracker"]
+__all__ = [
+    "VaultStreamCrawler",
+    "GraphQueryEngine",
+    "CrawlStateTracker",
+    "ContentIndex",
+]